Fixed bug #133
git-svn-id: file:///scratch/steve/svn/rapidSTORM-dev/rapidSTORM/trunk@3291 6dd997e4-92ef-4477-b1f0-deb9b3b8d6d5

+++ outputs/MemoryCache.cpp
@@ -49,7 +49,7 @@ struct StoreTree {
 
 class Localizations {
   protected:
-    short offset, count;
+    int offset, count;
     boost::optional< std::vector< Localizations > > children;
     typedef std::vector<Localization>::iterator Input;
     typedef std::vector<Localization>::const_iterator ConstInput;
